+++ app/commands/campaigns_context/join_campaign_command.rb
@@ -11,7 +11,8 @@ module CampaignsContext
       'Fate::Character' => 'fate',
       'Fallout::Character' => 'fallout',
       'Cosmere::Character' => 'cosmere',
-      'Cthulhu::Character' => 'cthulhu7'
+      'Cthulhu::Character' => 'cthulhu7',
+      'Nimble::Character' => 'nimble'
     }.freeze
 
     use_contract do