+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- campaigns for nimble characters
+
+### Modified
+- rendering dice roll results for dh characters
+
 ## [0.5.3] - 2026-07-30
 ### Added
 - filtering for spell classes for DC20 characters